+++ src/lib/types/index.ts
@@ -1,4 +1,5 @@
 // barrel file for types
 
 export * from "@/lib/types/course";
+export * from "@/lib/types/filters";
 export * from "@/lib/types/planner";
